Each engine now had its own thread. (THREAD SAFE!) The GameEngine thread dispatches the GraphicsEngine update routine. Since GL usage must be done from the same thread which GLFW context was initialized on.

+++ AIEngine/cAIObject.cpp
@@ -67,7 +67,7 @@ void cDalekAI::update(float deltaTime)
 	else
 	{
 		// Pick new target
-		this->pState->setCooldown((rand() % 6) + 1);
+		this->pState->setCooldown((float)(rand() % 6) + 1);
 
 		this->pState->setTarget(glm::vec3());
 	}
